[HW13] Create all class hibernatedao

+++ controller/src/main/java/com/senla/carservice/controller/CarOfficeController.java
@@ -37,8 +37,8 @@ public class CarOfficeController {
         }
         try {
             Date startDayDate = DateUtil.bringStartOfDayDate(dateFree);
-            int numberFreeMasters = masterService.getNumberFreeMastersByDate(startDayDate);
-            int numberFreePlace = placeService.getNumberFreePlaceByDate(startDayDate);
+            Long numberFreeMasters = masterService.getNumberFreeMastersByDate(startDayDate);
+            Long numberFreePlace = placeService.getNumberFreePlaceByDate(startDayDate);
             return "- number free places in service: " + numberFreePlace + "\n- number free masters in service: " +
                    numberFreeMasters;
         } catch (BusinessException | DaoException e) {
